cloud/requirements.txt
websockets>=12.0
openai-whisper
numpy
scipy
gtts
anthropic